requirements.txt
aiohttp==3.11.0
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
openai==1.58.1
propcache==0.2.1
requests==2.32.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
python-dotenv==1.1.1 